Rework faiserver setup.  Add instructions for the diskless server.
Create configuration for PXE boot dependent on the total number of
workstations. The number is fetched from the $WS_RANGE variable.
Diskless servers show a message how to install the chroot for the
diskless machines after setting up the FAI nfsroot.  Helps and
simplifies getting the system ready after installation.